% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_scores.R
\name{compute_ipw}
\alias{compute_ipw}
\title{Get inverse probability weighted (IPW) estimate from nuisance parameters}
\usage{
compute_ipw(data, g_output, e_output, idx_treat = NULL,
  idx_cntrl = NULL, shift_type = c("ipsi", "mtp"))
}
\arguments{
\item{data}{A \code{data.table} containing the observed data, with columns
in the order specified by the NPSEM (Y, Z, A, W), with column names set
appropriately based on the original input data. Such a structure is merely
a convenience utility to passing data around to the various core estimation
routines and is automatically generated as part of a call to the user-facing
wrapper function \code{medshift}.}

\item{g_output}{Object containing results from fitting the propensity score
regression, as produced by a call to \code{fit_g_mech}.}

\item{e_output}{Object containing results from fitting the propensity score
regression while conditioning on mediators, as produced by a call to
\code{fit_e_mech}.}

\item{idx_treat}{A \code{numeric} vector providing the indices corresponding
to units that received treatment (A = 1), for a binary intervention.}

\item{idx_cntrl}{A \code{numeric} vector providing the indices corresponding
to units that did not receive treatment (A = 0), for a binary intervention.}

\item{shift_type}{A choice of the type of stochastic treatment regime to use
-- either \code{"mtp"} for a modified treatment policy that shifts the
center of the observed intervention distribution by the scalar \code{delta}
or \code{"ipsi"} for an incremental propensity score shift that multiples
the odds of receiving the intervention by the scalar \code{delta}.}
}
\description{
Get inverse probability weighted (IPW) estimate from nuisance parameters
}
\keyword{internal}
