% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{confint.medshift}
\alias{confint.medshift}
\title{Confidence Intervals for Stochastic Mediation Parameters}
\usage{
\method{confint}{medshift}(object, parm = seq_len(object$psi), level = 0.95, ...)
}
\arguments{
\item{object}{An object of class \code{medshift}, as produced by invoking
\code{\link{medshift}}, for which a confidence interval is to be computed.}

\item{parm}{A \code{numeric} vector indicating indices of \code{object$est}
for which to return confidence intervals.}

\item{level}{A \code{numeric} indicating the confidence interval level.}

\item{...}{Other arguments. Not currently used.}
}
\description{
Compute confidence intervals for objects of class \code{medshift}, which
contain estimates produced by \code{\link{medshift}}.
}
