% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{confint.medshiftx}
\alias{confint.medshiftx}
\title{Confidence Intervals for Stochastic Mediation Parameters}
\usage{
\method{confint}{medshiftx}(object, parm = seq_len(object$psi),
  level = 0.95, ...)
}
\arguments{
\item{object}{An object of class \code{medshift}, as produced by invoking
the function \code{tmle_medshift}, for which a confidence interval is to be
computed.}

\item{parm}{A \code{numeric} vector indicating indices of \code{object$est}
for which to return confidence intervals.}

\item{level}{A \code{numeric} indicating the level of the confidence interval
to be computed.}

\item{...}{Other arguments. Not currently used.}
}
\description{
Compute confidence intervals for objects of class \code{medshiftx}, which
contain estimates produced by \code{medshift}.
}
