% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_eif.R
\name{cv_eif}
\alias{cv_eif}
\title{Cross-validated evaluation of EIF components}
\usage{
cv_eif(fold, data, delta, lrnr_stack_g, lrnr_stack_e, lrnr_stack_m,
  lrnr_stack_phi, z_names, w_names)
}
\arguments{
\item{fold}{Object specifying cross-validation folds as generated by a call
to \code{origami::make_folds}.}

\item{data}{A \code{data.table} containing the observed data, with columns
in the order specified by the NPSEM (Y, Z, A, W), with column names set
appropriately based on the original input data. Such a structure is merely
a convenience utility to passing data around to the various core estimation
routines and is automatically generated as part of a call to the user-facing
wrapper function \code{medshift}.}

\item{delta}{A \code{numeric} value indicating the degree of shift in the
intervention to be used in defining the causal quantity of interest. In the
case of binary interventions, this takes the form of an incremental
propensity score shift, acting as a multiplier of the probability with which
a given observational unit receives the intervention (EH Kennedy, 2018,
JASA; <doi:10.1080/01621459.2017.1422737>).}

\item{lrnr_stack_g}{A \code{Stack} object, or other learner class (inheriting
from \code{Lrnr_base}), containing a single or set of instantiated learners
from the \code{sl3} package, to be used in fitting a model for the
propensity score, i.e., g = P(A | W).}

\item{lrnr_stack_e}{A \code{Stack} object, or other learner class (inheriting
from \code{Lrnr_base}), containing a single or set of instantiated learners
from the \code{sl3} package, to be used in fitting a cleverly parameterized
propensity score that includes the mediators, i.e., e = P(A | Z, W).}

\item{lrnr_stack_m}{A \code{Stack} object, or other learner class (inheriting
from \code{Lrnr_base}), containing a single or set of instantiated learners
from the \code{sl3} package, to be used in fitting the outcome regression,
i.e., m(A, Z, W).}

\item{lrnr_stack_phi}{A \code{Stack} object, or other learner class
(inheriting from \code{Lrnr_base}), containing a single or set of
instantiated learners from the \code{sl3} package, to be used in fitting a
reduced regression useful for computing the efficient one-step estimator,
i.e., phi(W) = E[m(A = 1, Z, W) - m(A = 0, Z, W) | W).}

\item{z_names}{A \code{character} vector of the names of the columns that
correspond to mediators (Z). The input for this argument is automatically
generated by a call to the wrapper function \code{medshift}.}

\item{w_names}{A \code{character} vector of the names of the columns that
correspond to baseline covariates (W). The input for this argument is
automatically generated by a call to the wrapper function \code{medshift}.}
}
\description{
Cross-validated evaluation of EIF components
}
