% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_eff.R
\name{est_eff}
\alias{est_eff}
\title{Efficient estimator}
\usage{
est_eff(data, delta, g_lrnrs, e_lrnrs, m_lrnrs, phi_lrnrs, w_names,
  z_names)
}
\arguments{
\item{data}{...}

\item{delta}{...}

\item{g_lrnrs}{...}

\item{e_lrnrs}{...}

\item{m_lrnrs}{...}

\item{phi_lrnrs}{...}

\item{w_names}{...}

\item{z_names}{...}
}
\description{
Efficient estimator
}
\keyword{internal}
