% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_re.R
\name{est_re}
\alias{est_re}
\title{Re-weighted estimator}
\usage{
est_re(data, delta, g_lrnrs, e_lrnrs, w_names, z_names)
}
\arguments{
\item{data}{...}

\item{delta}{...}

\item{g_lrnrs}{...}

\item{e_lrnrs}{...}

\item{w_names}{...}

\item{z_names}{...}
}
\description{
Re-weighted estimator
}
\keyword{internal}
