% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_sub.R
\name{est_sub}
\alias{est_sub}
\title{Substitution estimator}
\usage{
est_sub(data, delta, g_lrnrs, m_lrnrs, w_names, z_names)
}
\arguments{
\item{data}{...}

\item{delta}{...}

\item{g_lrnrs}{...}

\item{m_lrnrs}{...}

\item{w_names}{...}

\item{z_names}{...}
}
\description{
Substitution estimator
}
\keyword{internal}
