% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_substitution.R
\name{est_substitution}
\alias{est_substitution}
\title{Substitution estimator}
\usage{
est_substitution(data, delta, g_learners, m_learners, w_names, z_names, ...)
}
\arguments{
\item{data}{A \code{data.table} containing the observed data, with columns
in the order specified by the NPSEM (Y, Z, A, W), with column names set
appropriately based on the original input data. Such a structure is merely
a convenience utility to passing data around to the various core estimation
routines and is automatically generated by \code{\link{medshift}}.}

\item{delta}{A \code{numeric} value indicating the degree of shift in the
intervention to be used in defining the causal quantity of interest. In the
case of binary interventions, this takes the form of an incremental
propensity score shift, acting as a multiplier of the odds with which a
given observational unit receives the intervention (EH Kennedy, 2018, JASA;
\url{doi:10.1080/01621459.2017.1422737}).}

\item{g_learners}{A \code{\link[sl3]{Stack}} (or other learner class that
inherits from \code{\link[sl3]{Lrnr_base}}), containing a single or set of
instantiated learners from \pkg{sl3}, to be used in fitting the propensity
score, i.e., g = P(A | W).}

\item{m_learners}{A \code{\link[sl3]{Stack}} (or other learner class that
inherits from \code{\link[sl3]{Lrnr_base}}), containing a single or set of
instantiated learners from \pkg{sl3}, to be used in fitting the outcome
regression, i.e., m(A, Z, W).}

\item{w_names}{A \code{character} vector of the names of the columns that
correspond to baseline covariates (W). The input for this argument is
automatically generated by a call to the wrapper function \code{medshift}.}

\item{z_names}{A \code{character} vector of the names of the columns that
correspond to mediators (Z). The input for this argument is automatically
generated by a call to the wrapper function \code{medshift}.}

\item{...}{Other arguments currently ignored.}
}
\description{
Substitution estimator
}
