% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_efficient.R
\name{estim_eff}
\alias{estim_eff}
\title{Efficient estimator}
\usage{
estim_eff(task, delta_shift, lrnr_stack_g, lrnr_stack_e, lrnr_stack_m)
}
\arguments{
\item{task}{...}

\item{delta_shift}{...}

\item{lrnr_stack_g}{...}

\item{lrnr_stack_e}{...}

\item{lrnr_stack_m}{...}
}
\description{
Efficient estimator
}
