% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_reweighted.R
\name{estim_re}
\alias{estim_re}
\title{Re-weighted estimator}
\usage{
estim_re(task, delta_shift, lrnr_stack_g, lrnr_stack_e)
}
\arguments{
\item{task}{...}

\item{delta_shift}{...}

\item{lrnr_stack_g}{...}

\item{lrnr_stack_e}{...}
}
\description{
Re-weighted estimator
}
