% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_reweighted.R
\name{estim_reweighted}
\alias{estim_reweighted}
\title{Re-weighted estimator}
\usage{
estim_reweighted(task, delta_shift, lrnr_stack)
}
\arguments{
\item{task}{...}

\item{delta_shift}{...}

\item{lrnr_stack}{...}
}
\description{
Re-weighted estimator
}
