% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_e_mech.R
\name{fit_e_mech}
\alias{fit_e_mech}
\title{Fit clever propensity score or intervention density for mediators}
\usage{
fit_e_mech(task, lrnr_stack)
}
\arguments{
\item{task}{...}

\item{lrnr_stack}{...}
}
\description{
Fit clever propensity score or intervention density for mediators
}
