% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_m_mech.R
\name{fit_m_mech}
\alias{fit_m_mech}
\title{Fit outcome regression}
\usage{
fit_m_mech(data, valid_data = NULL, lrnr_stack, z_names, w_names)
}
\arguments{
\item{data}{...}

\item{valid_data}{A validation/holdout data set (optional).}

\item{lrnr_stack}{...}

\item{z_names}{...}

\item{w_names}{...}
}
\description{
Fit outcome regression
}
