% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mechanisms.R
\name{fit_phi_mech}
\alias{fit_phi_mech}
\title{Fit intervention-specific exponential tilt nuisance parameter}
\usage{
fit_phi_mech(train_data, valid_data, learners, m_output, w_names)
}
\arguments{
\item{train_data}{A \code{data.table} containing the observed data, with
columns in the order specified by the NPSEM (Y, Z, A, W), with column names
set appropriately based on the input data. Such a structure is merely a
convenience utility to passing data around to the various core estimation
routines and is automatically generated by \code{\link{medshift}}.}

\item{valid_data}{A holdout data set, with columns exactly matching those
appearing in the preceding argument \code{train_data}, to be used for
estimation via cross-fitting. Not optional for this nuisance parameter.}

\item{learners}{A \code{\link[sl3]{Stack}} (or other learner class that
inherits from \code{\link[sl3]{Lrnr_base}}), containing a single or set of
instantiated learners from \pkg{sl3}, to be used in a regression of a
pseudo-outcome on the baseline covariates, i.e.,
phi(W) = E[m(A = 1, Z, W) - m(A = 0, Z, W) | W).}

\item{m_output}{Object containing results from fitting the outcome
regression, as produced by \code{\link{fit_m_mech}}.}

\item{w_names}{A \code{character} vector of the names of the columns that
correspond to baseline covariates (W). The input for this argument is
automatically generated by \code{\link{medshift}}.}
}
\description{
Fit intervention-specific exponential tilt nuisance parameter
}
