% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{integrate_over_g}
\alias{integrate_over_g}
\title{Numerical Integration Over Domain of Treatment Mechanism}
\usage{
integrate_over_g(g_mech, a_vals, weighting, int_grid_size = 10)
}
\arguments{
\item{g_mech}{...}

\item{a_vals}{...}

\item{weighting}{...}

\item{int_grid_size}{A \code{numeric} scalar corresponding to the number of
points to be used in numerically evaluating an integral over the domain of
the natural value of the treatment. This is only relevant in the case of
modified treatment policies for continuous-valued treatments; it is ignored
in the case of incremental propensity score interventions.}
}
\description{
In the case of modified treatment policies, it is necessary to numerically
evaluate an integral over the domain of the treatment mechanism. This is a
simple procedure to numerically compute such an integral based on Monte
Carlo importance sampling from a uniform distribution.
}
\keyword{internal}
