% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_interv.R
\name{interv_est_tmle}
\alias{interv_est_tmle}
\title{Compute Cross-Validated Targeted Minimum Loss-based Estimator (CV-TMLE)}
\usage{
interv_est_tmle(
  data,
  delta,
  g_learners,
  e_learners,
  b_learners,
  d_learners,
  m_learners,
  w_names,
  z_names,
  cv_folds = 5L,
  max_iter = 10L,
  tol_tilt = 10L
)
}
\arguments{
\item{data}{...}

\item{delta}{...}

\item{g_learners}{...}

\item{e_learners}{...}

\item{b_learners}{...}

\item{d_learners}{...}

\item{m_learners}{...}

\item{w_names}{A \code{character} vector of the names of the columns that
correspond to baseline covariates (W). The input for this argument is
automatically generated by a call to the wrapper function \code{medshift}.}

\item{z_names}{A \code{character} vector of the names of the columns that
correspond to mediators (Z). The input for this argument is automatically
generated by a call to the wrapper function \code{medshift}.}

\item{cv_folds}{...}

\item{max_iter}{...}

\item{tol_tilt}{...}
}
\description{
Compute Cross-Validated Targeted Minimum Loss-based Estimator (CV-TMLE)
}
