% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ipsi_delta}
\alias{ipsi_delta}
\title{Mapping for Incremental Propensity Score Interventions}
\usage{
ipsi_delta(g_est, delta)
}
\arguments{
\item{g_est}{A \code{numeric} vector of estimated propensity scores, ie, the
conditional probability of receiving treatment (A = 1) given covariates W.}

\item{delta}{A \code{numeric} value indicating the magnitude of shift in the
treatment that defines the counterfactual quantity of interest. In the case
of binary treatments, this defines the incremental propensity score shift,
which acts as a multiplier of the odds with which an observational unit
receives treatment (EH Kennedy, 2018; \url{doi:10.1080/01621459.2017.1422737}).
When the treatment is quantitative, the defines the degree of shift for a
modified treatment policy but non-binary treatments are not  yet supported.}
}
\description{
Mapping for Incremental Propensity Score Interventions
}
