% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_mechanisms.R
\name{ipsi_shift}
\alias{ipsi_shift}
\title{Incremental Propensity Score Intervention}
\usage{
ipsi_shift(gn_est, delta)
}
\arguments{
\item{gn_est}{A \code{numeric} vector of the estimated propensity score, the
probability of receiving treatment A given baseline covariates W.}

\item{delta}{A \code{numeric} scalar giving the degree by which the odds of
receiving treatment are shifted in this type of intervention.}
}
\description{
Compute the exponential tilt shift, which corresponds to an incremental
propensity score intervention on gn := Pr(A = 1 | W).
}
