% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LF_exptilt_ipsi.R
\docType{data}
\name{LF_exptilt_ipsi}
\alias{LF_exptilt_ipsi}
\title{Exponential Tilt Likelihood Factor for Incremental Propensity Scores}
\format{\code{\link{R6Class}} object.}
\usage{
LF_exptilt_ipsi
}
\value{
\code{LF_base} object
}
\description{
...
}
\section{Constructor}{

  \code{define_lf(LF_exptilt_ipsi, name, type = "density", likelihood_base,
    shift_param, ...)}

  \describe{
    \item{\code{name}}{character, the name of the factor. Should match a node
          name in the nodes specified by \code{\link{tmle3_Task}$npsem}.
    }
    \item{\code{likelihood_base}}{The trained \code{\link{likelihood}}
          object, for use in generating a re-scaled likelihood factor.
    }
    \item{\code{shift_param}}{\code{numeric}, specification of the magnitude
          of the desired shift (a multiplier for the propensity score).
    }
    \item{\code{...}}{Not currently used.
    }
  }
}

\section{Fields}{

\describe{
    \item{\code{likelihood_base}}{The trained \code{\link{likelihood}}
          object, for use in generating a re-scaled likelihood factor.
    }
    \item{\code{shift_param}}{\code{numeric}, specification of the magnitude
          of the desired shift (a multiplier for the propensity score).
    }
    \item{\code{...}}{Additional arguments passed to the base class.
    }
  }
}

\references{
\describe{
  \item{"Nonparametric Causal Effects Based on Incremental Propensity Score
        Interventions."}{Edward's JASA paper...}
  \item{"Causal Mediation Analysis for Stochastic Interventions"}{Díaz, Iván
        and Hejazi, Nima S (2019). submitted.}
}
}
\concept{Likelihood objects}
\keyword{data}
