% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LF_ipsi.R
\name{LF_ipsi}
\alias{LF_ipsi}
\title{Likelihood Factor for Incremental Propensity Score Interventions}
\format{
\code{\link[R6]{R6Class}} object.
}
\value{
\code{\link[tmle3]{LF_base}} object.
}
\description{
Likelihood Factor for Incremental Propensity Score Interventions
}
\section{Constructor}{

\code{define_lf(LF_ipsi, name, type = "density", likelihood_base,
                  shift_param, treatment_task, control_task, ...)}

\describe{
\item{\code{name}}{A \code{character}, giving the name of the likelihood
factor. Should match a node name in the nodes specified by the
\code{npsem} slot of \code{\link[tmle3]{tmle3_Task}}.}
\item{\code{likelihood_base}}{A trained \code{\link[tmle3]{Likelihood}}
object, for use in generating a re-scaled likelihood factor.}
\item{\code{shift_param}}{A \code{numeric}, specifying the magnitude of
the desired incremental propensity score shift (a multiplier of
the odds of receiving treatment).}
\item{\code{treatment_task}}{A \code{\link[tmle3]{tmle3_Task}} object
created by setting the intervention to the treatment condition:
do(A = 1).}
\item{\code{control_task}}{A \code{\link[tmle3]{tmle3_Task}} object
created by setting the intervention to the control condition:
do(A = 0).}
\item{\code{...}}{Not currently used.}
}
}

\section{Fields}{

\describe{
\item{\code{likelihood_base}}{A trained \code{\link[tmle3]{Likelihood}}
object, for use in generating a re-scaled likelihood factor.
}
\item{\code{shift_param}}{A \code{numeric}, specifying the magnitude of
the desired incremental propensity score shift (a multiplier of
the odds of receiving treatment).}
\item{\code{treatment_task}}{A \code{\link[tmle3]{tmle3_Task}} object
created by setting the intervention to the treatment condition:
do(A = 1).}
\item{\code{control_task}}{A \code{\link[tmle3]{tmle3_Task}} object
created by setting the intervention to the control condition:
do(A = 0).}
\item{\code{...}}{Additional arguments passed to the base class.}
}
}

\references{
\describe{
\item{"Nonparametric Causal Effects Based on Incremental Propensity Score
Interventions."}{Kennedy, Edward H (2019). Journal of the American
Statistical Association.
\url{https://doi.org/10.1080/01621459.2017.1422737}}
\item{"Causal Mediation Analysis for Stochastic Interventions"}{Díaz,
Iván and Hejazi, Nima S (2020). Journal of the Royal Statistical
Society, Series B. \url{https://doi.org/10.1111/rssb.12362}}
}
}
\concept{Likelihood objects}
\keyword{data}
