% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_utils.R
\name{make_Dzw}
\alias{make_Dzw}
\title{Compute Dzw component of efficient influence funtion}
\usage{
make_Dzw(g_output, m_output)
}
\arguments{
\item{g_output}{...}

\item{m_output}{...}
}
\description{
Compute Dzw component of efficient influence funtion
}
