% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle3_helpers.R
\name{make_e_task}
\alias{make_e_task}
\title{Make task for derived likelihood factor e(A,W)}
\usage{
make_e_task(tmle_task, likelihood)
}
\arguments{
\item{tmle_task}{...}

\item{likelihood}{...}
}
\description{
Make task for derived likelihood factor e(A,W)
}
