% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle3_helpers.R
\name{make_phi_task}
\alias{make_phi_task}
\title{Make task for derived likelihood factor phi(W)}
\usage{
make_phi_task(tmle_task, likelihood)
}
\arguments{
\item{tmle_task}{...}

\item{likelihood}{...}
}
\description{
Make task for derived likelihood factor phi(W)
}
