% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mc_integrate_dens}
\alias{mc_integrate_dens}
\title{Numerical Integration for Weighted Treatment Mechanism}
\usage{
mc_integrate_dens(data, delta, mc_draws, dens_mech, wts_mech,
  shift_type = c("mtp", "ipsi"))
}
\arguments{
\item{data}{A \code{data.table} containing the observed data, with columns
in the order specified by the NPSEM (Y, Z, A, W), with column names set
appropriately based on the original input data. Such a structure is merely
a convenience utility to passing data around to the various core estimation
routines and is automatically generated as part of a call to the user-facing
wrapper function \code{\link{medshift}}.}

\item{delta}{A \code{numeric} value indicating the degree of shift in the
intervention to be used in defining the causal quantity of interest. In the
case of binary interventions, this takes the form of an incremental
propensity score shift, acting as a multiplier of the probability with which
a given observational unit receives the intervention (EH Kennedy, 2018,
JASA; <doi:10.1080/01621459.2017.1422737>). In the case of continuous
interventions, this is a modified treatment policy that shifts the observed
treatment by the given value.}

\item{mc_draws}{A \code{numeric} vector corresponding to draws from a uniform
distribution based on the range of the observed treatment mechanism. The
numerical integral is computed by using this grid as values of intervention
mechanism.}

\item{dens_mech}{Estimated conditional density corresponding to the natural
or shifted value of the treatment for modified treatment policies based on
the observed values of the treatment. This is an object with inherited class
\code{Lrnr_base}, automatically produced by \code{\link{fit_g_mech}}.}

\item{wts_mech}{Estimated weighting mechanism learned from the observed data
For the substitution estimator, this is the outcome mechanism, while, for
the intervention score, this is a nuisance parameter defined by the type
of shift intervention to be evaluated. This is an object with inherited
class \code{Lrnr_base}.}

\item{shift_type}{A choice of the type of stochastic treatment regime to use
-- either \code{"mtp"} for a modified treatment policy that shifts the
center of the observed intervention distribution by the scalar \code{delta}
or \code{"ipsi"} for an incremental propensity score shift that multiples
the odds of receiving the intervention by the scalar \code{delta}.}
}
\description{
In the case of modified treatment policies, it is necessary to numerically
evaluate an integral over the domain of the treatment mechanism. This is a
simple procedure to numerically compute such an integral based on Monte
Carlo importance sampling from a uniform distribution.
}
\keyword{internal}
