% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_mechanisms.R
\name{mtp_shift_inv}
\alias{mtp_shift_inv}
\title{Inverse Additive Modified Treatment Policy}
\usage{
mtp_shift_inv(A, W = NULL, delta)
}
\arguments{
\item{A}{The \code{numeric} vector of natural observed values of treatment.}

\item{W}{The \code{numeric} vector, \code{matrix}, \code{data.frame} or
similar object giving the values of the baseline covariates.}

\item{delta}{The \code{numeric} scalar giving the degree by which naturally
observed values of the intervention \code{A} ought to be shifted.}
}
\description{
Compute the inverse shift intervention: d(A,W) = A - delta, a term appearing
in the definition of the shifted intervention quantity for this shift type.
}
