% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Param_binary_medshift.R
\docType{data}
\name{Param_esteqn_medshift}
\alias{Param_esteqn_medshift}
\title{Parameter for the pure mediated effect under binary stochastic interventions}
\format{\code{\link{R6Class}} object.}
\usage{
Param_esteqn_medshift
}
\value{
\code{Param_base} object
}
\description{
Parameter definition...
}
\section{Constructor}{

  \code{define_param(Param_esteqn_medshift, observed_likelihood, intervention_list, ..., outcome_node)}

  \describe{
    \item{\code{observed_likelihood}}{A \code{\link{Likelihood}} corresponding to the observed likelihood
    }
    \item{\code{intervention_list}}{A list of objects inheriting from \code{\link{LF_base}}, representing the intervention.
    }
    \item{\code{...}}{Not currently used.
    }
    \item{\code{outcome_node}}{character, the name of the node that should be treated as the outcome
    }
    }
}

\section{Fields}{

\describe{
    \item{\code{cf_likelihood}}{the counterfactual likelihood for this treatment
    }
    \item{\code{intervention_list}}{A list of objects inheriting from \code{\link{LF_base}}, representing the intervention
    }
}
}

\concept{Parameters}
\keyword{data}
