% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Param_medshift.R
\name{Param_medshift}
\alias{Param_medshift}
\title{Parameter for the Population Intervention (In)direct Effects}
\format{
\code{\link[R6]{R6Class}} object.
}
\value{
\code{\link[tmle3]{Param_base}} object.
}
\description{
Parameter definition class. See \url{https://doi.org/10.1111/rssb.12362}.
}
\section{Constructor}{

\code{define_param(Param_medshift, shift_param, ..., outcome_node)}

\describe{
\item{\code{observed_likelihood}}{A \code{\link[tmle3]{Likelihood}}
corresponding to the observed likelihood.}
\item{\code{shift_param}}{A \code{numeric}, specifying the magnitude of
the desired incremental propensity score shift (a multiplier of
the odds of receiving treatment).}
\item{\code{...}}{Not currently used.}
\item{\code{outcome_node}}{A \code{character}, giving the name of the
node that should be treated as the outcome.}
}
}

\section{Fields}{

\describe{
\item{\code{cf_likelihood}}{The counterfactual likelihood under the
joint stochastic intervention on exposure and mediators.}
\item{\code{lf_ipsi}}{Object derived from \code{\link[tmle3]{LF_base}}
for assessing the joint intervention on exposure and mediators.}
\item{\code{treatment_task}}{A \code{\link[tmle3]{tmle3_Task}} created
by setting the intervention to the treatment condition:
do(A = 1).}
\item{\code{control_task}}{A \code{\link{tmle3_Task}} object created by
setting the intervention to the control condition: do(A = 0).}
\item{\code{shift_param}}{A \code{numeric}, specifying the magnitude of
the desired incremental propensity score shift (a multiplier of
the odds of receiving treatment).}
}
}

\concept{Parameters}
\keyword{data}
