% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pide.R
\name{pide}
\alias{pide}
\title{One-step or TML estimation of the population intervention direct effect}
\usage{
pide(
  W,
  A,
  Z,
  Y,
  ids = seq(1, length(Y)),
  delta,
  estimator = c("onestep", "tmle"),
  ci_level = 0.95,
  ...
)
}
\arguments{
\item{W}{A \code{matrix}, \code{data.frame}, or similar corresponding to a
set of baseline covariates.}

\item{A}{A \code{numeric} vector corresponding to a treatment variable. The
parameter of interest is defined as a location shift of this quantity.}

\item{Z}{A \code{numeric} vector, \code{matrix}, \code{data.frame}, or
similar corresponding to a set of mediators (on the causal pathway between
the intervention A and the outcome Y).}

\item{Y}{A \code{numeric} vector corresponding to an outcome variable.}

\item{ids}{A \code{numeric} vector of observation-level IDs, allowing for
observational units to be related through a hierarchical structure. The
default is to assume all units are IID. When repeated IDs are included,
both the cross-validation procedures used for estimation and inferential
procedures respect these IDs.}

\item{delta}{A \code{numeric} value indicating the degree of shift in the
intervention to be used in defining the causal quantity of interest. In the
case of binary interventions, this takes the form of an incremental
propensity score shift, acting as a multiplier of the odds with which a
unit receives the intervention (EH Kennedy, 2018, JASA;
\url{doi:10.1080/01621459.2017.1422737}).}

\item{estimator}{The desired estimator of the natural direct effect to be
computed. Currently, choices are limited to a substitution estimator, a
re-weighted estimator, a one-step estimator, and a targeted minimum loss
estimator.}

\item{ci_level}{A \code{numeric} indicating the desired coverage level of
the confidence interval to be computed.}

\item{...}{Additional arguments passed to \code{\link{medshift}}. Consult
the documentation of that function for details.}
}
\description{
One-step or TML estimation of the population intervention direct effect
}
