% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scale_from_unit}
\alias{scale_from_unit}
\title{Map values from the unit interval to their original scale}
\usage{
scale_from_unit(vals_unit)
}
\arguments{
\item{vals_unit}{A \code{numeric} vector of values scaled to fall in the
closed unit interval \link{0, 1} by use of \code{\link{scale_to_unit}}.}
}
\description{
Map values from the unit interval to their original scale
}
\keyword{internal}
