% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_stoch.R
\name{stoch_est_onestep}
\alias{stoch_est_onestep}
\title{Efficient One-Step Estimator}
\usage{
stoch_est_onestep(
  data,
  delta,
  g_learners,
  e_learners,
  m_learners,
  phi_learners,
  w_names,
  z_names,
  cv_folds = 10
)
}
\arguments{
\item{data}{A \code{data.table} containing the observed data, with columns
in the order specified by the NPSEM (Y, Z, A, W), with column names set
appropriately based on the original input data. Such a structure is merely
a convenience utility to passing data around to the various core estimation
routines and is automatically generated by \code{\link{medshift}}.}

\item{delta}{A \code{numeric} value indicating the magnitude of shift in the
treatment that defines the counterfactual quantity of interest. In the case
of binary treatments, this defines the incremental propensity score shift,
which acts as a multiplier of the odds with which an observational unit
receives treatment (EH Kennedy, 2018; \url{doi:10.1080/01621459.2017.1422737}).
When the treatment is quantitative, the defines the degree of shift for a
modified treatment policy but non-binary treatments are not  yet supported.}

\item{g_learners}{A \code{\link[sl3]{Stack}} (or other learner class that
inherits from \code{\link[sl3]{Lrnr_base}}), containing a single or set of
instantiated learners from \pkg{sl3}, used to fit a propensity score model.}

\item{e_learners}{A \code{\link[sl3]{Stack}} (or other learner class that
inherits from \code{\link[sl3]{Lrnr_base}}), containing a single or set of
instantiated learners from \pkg{sl3}, used to fit a reparametrized model
for the mediator(s) that is equivalent to a propensity score conditioning
on the mediators.}

\item{m_learners}{A \code{\link[sl3]{Stack}} (or other learner class that
inherits from \code{\link[sl3]{Lrnr_base}}), containing a single or set of
instantiated learners from \pkg{sl3}, used to fit an outcome regression.}

\item{phi_learners}{A \code{\link[sl3]{Stack}} (or other learner class that
inherits from \code{\link[sl3]{Lrnr_base}}), containing a single or set of
instantiated learners from \pkg{sl3}, used to fit a nuisance regression of
a "derived" pseudo-outcome (contrast of outcome regressions under different
treatment conditions), conditional on baseline covariates. This is of the
following form phi(W) = E[m(A = 1, Z, W) - m(A = 0, Z, W) | W).}

\item{w_names}{A \code{character} vector of the names of the columns that
correspond to baseline covariates (W). The input for this argument is
automatically generated by \code{\link{medshift}}.}

\item{z_names}{A \code{character} vector of the names of the columns that
correspond to mediators (Z). The input for this argument is automatically
generated by \code{\link{medshift}}.}

\item{cv_folds}{A \code{numeric} specifying the number of folds to be
created for cross-validation. Use of cross-validation / cross-fitting
allows for entropy conditions on the efficient estimator to be relaxed. For
compatibility with \code{\link[origami]{make_folds}}, this value must be
greater than or equal to 2; the default is to create 10 folds.}
}
\description{
Efficient One-Step Estimator
}
