% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle3_Spec_medshift.R
\name{stochastic_mediation_likelihood}
\alias{stochastic_mediation_likelihood}
\title{Stochastic Mediation Likelihood Factors}
\usage{
stochastic_mediation_likelihood(tmle_task, learner_list)
}
\arguments{
\item{tmle_task}{A \code{\link[tmle3]{tmle3_Task}} specifying the data and
NPSEM for use in constructing components required for TML estimation.}

\item{likelihood}{A trained \code{\link[tmle3]{Likelihood}}, constructed via
the \code{\link{stochastic_mediation_likelihood}} helper.}
}
\description{
Stochastic Mediation Likelihood Factors
}
\keyword{internal}
