% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_de.R
\name{test_de}
\alias{test_de}
\title{Hypothesis test of direct effect with mediated stochastic interventions
using the multiplier bootstrap}
\usage{
test_de(W, A, Z, Y, delta_grid = seq(from = 0, to = 1, by = 0.2),
  n_mult = 10000, mult_type = c("rademacher", "gaussian"),
  ci_level = 0.95, ...)
}
\arguments{
\item{W}{A \code{matrix}, \code{data.frame}, or similar corresponding to a
set of baseline covariates.}

\item{A}{A \code{numeric} vector corresponding to a treatment variable. The
parameter of interest is defined as a location shift of this quantity.}

\item{Z}{A \code{numeric} vector, \code{matrix}, \code{data.frame}, or
similar corresponding to a set of mediators (on the causal pathway between
the intervention A and the outcome Y).}

\item{Y}{A \code{numeric} vector corresponding to an outcome variable.}

\item{n_mult}{A \code{numeric} scalar giving the number of repetitions of the
multipliers (Rademacher or Gaussian) to be used in computing the multiplier
bootstrap.}

\item{mult_type}{A \code{character} identifying the type of multipliers to be
used in the multiplier bootstrap. Choices are limited to \code{"rademacher"}
or \code{"gaussian"}, with the default being the former.}

\item{...}{Other arguments to be passed to \code{\link{medshift}}.}

\item{delta}{A \code{numeric} value indicating the degree of shift in the
intervention to be used in defining the causal quantity of interest. In the
case of binary interventions, this takes the form of an incremental
propensity score shift, acting as a multiplier of the probability with which
a given observational unit receives the intervention (EH Kennedy, 2018,
JASA; <doi:10.1080/01621459.2017.1422737>).}

\item{error_level}{A \code{numeric} [TO FILL IN]}
}
\description{
Hypothesis test of direct effect with mediated stochastic interventions
using the multiplier bootstrap
}
