% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spec_esteqn_medshift.R
\docType{data}
\name{tmle3_Spec_medshift}
\alias{tmle3_Spec_medshift}
\title{Defines a TML Estimator for the Outcome under a Shifted Treatment}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
tmle3_Spec_medshift
}
\description{
Current limitations: pretty much tailored to \code{Param_TSM}
See TODO notes for places generalization can be added
}
\keyword{datasets}
