% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_utils.R
\name{compute_Dzw}
\alias{compute_Dzw}
\title{Get Dzw component of efficient influence function from nuisance parameters}
\usage{
compute_Dzw(g_output, m_output)
}
\arguments{
\item{g_output}{...}

\item{m_output}{...}
}
\description{
Get Dzw component of efficient influence function from nuisance parameters
}
