% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_stoch.R
\name{compute_Dzw}
\alias{compute_Dzw}
\title{Construct joint Z-W component of efficient influence function}
\usage{
compute_Dzw(g_out, m_out)
}
\arguments{
\item{g_out}{Object containing results from fitting the propensity score
regression, as produced by a call to \code{\link{fit_g_mech}}.}

\item{m_out}{Object containing results from fitting the outcome
regression, as produced by a call to \code{\link{fit_m_mech}}.}
}
\description{
Construct joint Z-W component of efficient influence function
}
\keyword{internal}
