% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_ipw.R
\name{compute_ipw}
\alias{compute_ipw}
\title{Get inverse probability weighted (IPW) estimate from nuisance parameters}
\usage{
compute_ipw(g_output, e_output, idx_treat, idx_cntrl)
}
\arguments{
\item{g_output}{Object containing results from fitting the propensity score
regression, as produced by a call to \code{\link{fit_g_mech}}.}

\item{e_output}{Object containing results from fitting the propensity score
regression while conditioning on mediators, as produced by a call to
\code{\link{fit_e_mech}}.}

\item{idx_treat}{A \code{numeric} vector providing the indices corresponding
to units that received treatment (A = 1), for a binary intervention.}

\item{idx_cntrl}{A \code{numeric} vector providing the indices corresponding
to units that did not receive treatment (A = 0), for a binary intervention.}
}
\description{
Get inverse probability weighted (IPW) estimate from nuisance parameters
}
\keyword{internal}
