% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_utils.R
\name{compute_ipw}
\alias{compute_ipw}
\title{Get inverse probability weighted (IPW) estimate from nuisance parameters}
\usage{
compute_ipw(g_output, e_output, idx_treat, idx_cntrl)
}
\arguments{
\item{g_output}{...}

\item{e_output}{...}

\item{idx_treat}{...}

\item{idx_cntrl}{...}
}
\description{
Get inverse probability weighted (IPW) estimate from nuisance parameters
}
