% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_eif.R
\name{cv_eif}
\alias{cv_eif}
\title{Cross-validated evaluation of EIF components}
\usage{
cv_eif(fold, data, delta_shift, lrnr_stack_g, lrnr_stack_e, lrnr_stack_m)
}
\arguments{
\item{fold}{Object specifying cross-validation folds as generated by a call
to \code{origami::make_folds}.}

\item{data}{...}

\item{delta_shift}{...}

\item{lrnr_stack_g}{...}

\item{lrnr_stack_e}{...}

\item{lrnr_stack_m}{...}
}
\description{
Cross-validated evaluation of EIF components
}
