% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_e_mech.R
\name{fit_e_mech}
\alias{fit_e_mech}
\title{Fit clever propensity score regression for mediators}
\usage{
fit_e_mech(data, lrnr_stack, z_names, w_names)
}
\arguments{
\item{data}{...}

\item{lrnr_stack}{...}

\item{z_names}{...}

\item{w_names}{...}
}
\description{
Fit clever propensity score regression for mediators
}
