% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_nuisance_stoch.R
\name{fit_e_mech}
\alias{fit_e_mech}
\title{Fit reparametrized propensity score conditioning on mediators}
\usage{
fit_e_mech(data, valid_data = NULL, learners, z_names, w_names)
}
\arguments{
\item{data}{A \code{data.table} containing the observed data, with columns
in the order specified by the NPSEM (Y, Z, L, A, W), with column names set
appropriately based on the original input data. Such a structure is merely
a convenience utility to passing data around to the various core estimation
routines and is automatically generated by \code{\link{medshift}}.}

\item{valid_data}{A holdout data set, with columns exactly matching those
appearing in the preceding argument \code{data}, to be used for estimation
via cross-fitting. Optional, defaulting to \code{NULL}.}

\item{learners}{A \code{\link[sl3]{Stack}} (or other learner class that
inherits from \code{\link[sl3]{Lrnr_base}}), containing a single or set of
instantiated learners from \pkg{sl3}, used to fit a reparametrized model
for the mediator(s) that is equivalent to a propensity score conditioning
on the mediators.}

\item{z_names}{A \code{character} vector of the names of the columns that
correspond to mediators (Z). The input for this argument is automatically
generated by \code{\link{medshift}}.}

\item{w_names}{A \code{character} vector of the names of the columns that
correspond to baseline covariates (W). The input for this argument is
automatically generated by a call to the wrapper function \code{medshift}.}
}
\description{
Fit reparametrized propensity score conditioning on mediators
}
