% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_g_mech.R
\name{fit_g_mech}
\alias{fit_g_mech}
\title{Fit propensity score with incremental shift}
\usage{
fit_g_mech(data, delta_shift, lrnr_stack, w_names)
}
\arguments{
\item{data}{...}

\item{delta_shift}{...}

\item{lrnr_stack}{...}

\item{w_names}{...}
}
\description{
Fit propensity score with incremental shift
}
