% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_g_mech.R
\name{fit_g_mech}
\alias{fit_g_mech}
\title{Fit propensity score or intervention density}
\usage{
fit_g_mech(task, delta_shift, lrnr_stack)
}
\arguments{
\item{task}{...}

\item{delta_shift}{...}

\item{lrnr_stack}{...}
}
\description{
Fit propensity score or intervention density
}
