% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_interv.R
\name{interv_est_onestep}
\alias{interv_est_onestep}
\title{Compute Cross-Fitted One-Step Estimator}
\usage{
interv_est_onestep(
  data,
  delta,
  g_learners,
  e_learners,
  b_learners,
  d_learners,
  m_learners,
  w_names,
  z_names,
  cv_folds = 5L
)
}
\arguments{
\item{data}{...}

\item{delta}{...}

\item{g_learners}{...}

\item{e_learners}{...}

\item{b_learners}{...}

\item{d_learners}{...}

\item{m_learners}{...}

\item{cv_folds}{...}
}
\description{
Compute Cross-Fitted One-Step Estimator
}
