% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_interv.R
\name{interv_est_tmle}
\alias{interv_est_tmle}
\title{Compute Cross-Validated TML estimator}
\usage{
interv_est_tmle(
  data,
  delta,
  g_learners,
  e_learners,
  b_learners,
  d_learners,
  m_learners,
  cv_folds = 5L,
  tiltmod_tol = 10
)
}
\arguments{
\item{data}{...}

\item{delta}{...}

\item{g_learners}{...}

\item{e_learners}{...}

\item{b_learners}{...}

\item{d_learners}{...}

\item{m_learners}{...}

\item{cv_folds}{...}

\item{tiltmod_tol}{...}
}
\description{
Compute Cross-Validated TML estimator
}
