% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LF_exptilt_ipsi.R
\docType{data}
\name{LF_exptilt_ipsi}
\alias{LF_exptilt_ipsi}
\title{Exponential Tilt Likelihood Factor for Incremental Propensity Scores}
\format{\code{R6::R6Class} object.}
\usage{
LF_exptilt_ipsi
}
\value{
\code{tmle3::LF_base} object
}
\description{
Exponential Tilt Likelihood Factor for Incremental Propensity Scores
}
\section{Constructor}{

  \code{define_lf(LF_exptilt_ipsi, name, type = "density", likelihood_base,
    shift_param, treatment_task, control_task, ...)}

  \describe{
    \item{\code{name}}{character, the name of the factor. Should match a node
          name in the nodes specified by \code{tmle3::tmle3_Task$npsem}.
    }
    \item{\code{likelihood_base}}{The trained \code{tmle3::Likelihood}
          object, for use in generating a re-scaled likelihood factor.
    }
    \item{\code{shift_param}}{\code{numeric}, specification of the magnitude
          of the desired shift (a multiplier for the propensity score).
    }
    \item{\code{treatment_task}}{\code{tmle3::tmle3_Task} object created from
          setting the intervention to the treatment condition: do(A = 1).
    }
    \item{\code{control_task}}{\code{tmle3::tmle3_Task} object created from
          setting the intervention to the control condition: do(A = 0).
    }
    \item{\code{...}}{Not currently used.
    }
  }
}

\section{Fields}{

\describe{
    \item{\code{likelihood_base}}{The trained \code{tmle3::Likelihood}
          object, for use in generating a re-scaled likelihood factor.
    }
    \item{\code{shift_param}}{\code{numeric}, specification of the magnitude
          of the desired shift (a multiplier for the propensity score).
    }
    \item{\code{treatment_task}}{\code{tmle3::tmle3_Task} object created from
          setting the intervention to the treatment condition: do(A = 1).
    }
    \item{\code{control_task}}{\code{tmle3::tmle3_Task} object created from
          setting the intervention to the control condition: do(A = 0).
    }
    \item{\code{...}}{Additional arguments passed to the base class.
    }
  }
}

\references{
\describe{
  \item{"Nonparametric Causal Effects Based on Incremental Propensity Score
        Interventions."}{Edward's JASA paper...}
  \item{"Causal Mediation Analysis for Stochastic Interventions"}{Díaz, Iván
        and Hejazi, Nima S (2019). submitted.}
}
}
\concept{Likelihood objects}
\keyword{data}
