% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medshift.R
\name{medshift}
\alias{medshift}
\title{Nonparametric estimation of direct effects under mediation}
\usage{
medshift(W, A, Z, Y, shift_value = 0.5,
  g_lrnrs = Lrnr_glm_fast$new(family = binomial()),
  e_lrnrs = Lrnr_glm_fast$new(family = binomial()),
  m_lrnrs = Lrnr_glm_fast$new(), estimator = c("efficient",
  "substitution", "reweighted"))
}
\arguments{
\item{W}{...}

\item{A}{...}

\item{Z}{...}

\item{Y}{...}

\item{shift_value}{...}

\item{g_lrnrs}{...}

\item{e_lrnrs}{...}

\item{estimator}{...}

\item{Q_lrnrs}{...}
}
\description{
Nonparametric estimation of direct effects under mediation
}
