% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medshift.R
\name{medshift}
\alias{medshift}
\title{Nonparametric estimation of direct effects under mediation}
\usage{
medshift(W, A, Z, Y, g_lrnrs, e_lrnrs, Q_lrnrs)
}
\arguments{
\item{W}{...}

\item{A}{...}

\item{Z}{...}

\item{Y}{...}

\item{g_lrnrs}{...}

\item{e_lrnrs}{...}

\item{Q_lrnrs}{...}
}
\description{
Nonparametric estimation of direct effects under mediation
}
