% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Param_stochastic_mediation.R
\docType{data}
\name{Param_medshift}
\alias{Param_medshift}
\title{Parameter for mediation effect under stochastic interventions}
\format{\code{\link{R6Class}} object.}
\usage{
Param_medshift
}
\value{
\code{Param_base} object
}
\description{
Parameter definition...
}
\section{Constructor}{

  \code{define_param(Param_esteqn_medshift, observed_likelihood, intervention_list, ..., outcome_node)}

  \describe{
    \item{\code{observed_likelihood}}{A \code{\link{Likelihood}}
          corresponding to the observed likelihood.
    }
    \item{\code{shift_param}}{\code{numeric}, specification of the magnitude
          of the desired shift (a multiplier for the propensity score).
    }
    \item{\code{...}}{Not currently used.
    }
    \item{\code{outcome_node}}{character, the name of the node that should be
          treated as the outcome
    }
  }
}

\section{Fields}{

\describe{
    \item{\code{cf_likelihood}}{the counterfactual likelihood under the
          stochastic intervention on mediators and treatment.
    }
}
}

\concept{Parameters}
\keyword{data}
