% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle3_helpers.R
\name{stochastic_mediation_npsem}
\alias{stochastic_mediation_npsem}
\title{Stochastic Mediation NPSEM}
\usage{
stochastic_mediation_npsem(node_list, variable_types = NULL)
}
\arguments{
\item{node_list}{A \code{list} object specifying the different nodes in the
nonparametric structural equation model.}

\item{variable_types}{Used to define how variables are handled. No need for
user specification of this argument.}
}
\description{
Stochastic Mediation NPSEM
}
