% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle3_Spec_medshift.R
\name{stochastic_mediation_npsem}
\alias{stochastic_mediation_npsem}
\title{Stochastic Mediation NPSEM}
\usage{
stochastic_mediation_npsem(node_list, variable_types = NULL)
}
\arguments{
\item{node_list}{A \code{list} object specifying the different nodes in the
nonparametric structural equation model (NPSEM).}

\item{variable_types}{Used to define how variables are handled. Optional.}
}
\description{
Stochastic Mediation NPSEM
}
\keyword{internal}
