% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle3_Spec_medshift.R
\name{tmle3_Spec_medshift}
\alias{tmle3_Spec_medshift}
\title{Defines a TML Estimator for the Counterfactual Mean Under Joint Stochastic
Intervention on Treatment and Mediator}
\description{
Defines a TML Estimator for the Counterfactual Mean Under Joint Stochastic
Intervention on Treatment and Mediator
}
